# Speed up OSMData -> MapData